requirements.txt
#------------------------------------------------------------------------------
#--- !!!   IMPORTANT   !!!
#    This is a Collection of all Project Dependencies.
#    !!! DO NOT REMOVE !!!
#------------------------------------------------------------------------------
#--- Upgrade all
#    pip freeze --local | grep -v "^\-e" | cut -d = -f 1 | xargs -n1 pip install -U

#------------------------------------------------------------------------------
# --- "#"
Django==4.2.11

#------------------------------------------------------------------------------
# --- "A"

# -----------------------------------------------------------------------------
# --- "B"

# -----------------------------------------------------------------------------
# --- "C"
coverage==7.2.7

# -----------------------------------------------------------------------------
# --- "D"
djangorestframework==3.15.1

# -----------------------------------------------------------------------------
# --- "django-*"

#------------------------------------------------------------------------------
# --- "E"

#------------------------------------------------------------------------------
# --- "F"

#------------------------------------------------------------------------------
# --- "G"
gunicorn==20.1.0

#------------------------------------------------------------------------------
# --- "H"

#------------------------------------------------------------------------------
# --- "I"
ipython==8.18.1

#------------------------------------------------------------------------------
# --- "J"

#------------------------------------------------------------------------------
# --- "K"

#------------------------------------------------------------------------------
# --- "L"

#------------------------------------------------------------------------------
# --- "M"
markdown==3.4.3
mock==5.1.0

#------------------------------------------------------------------------------
# --- "N"

#------------------------------------------------------------------------------
# --- "O"

#------------------------------------------------------------------------------
# --- "P"
pep8==1.7.1
pep257==0.7.0
psycopg2-binary==2.9.5
pycodestyle==2.5.0
pylint==2.15.10
pylint-django==2.5.3
pytest==7.4.0

# -----------------------------------------------------------------------------
# --- "python-*"
python-decouple==3.8

#------------------------------------------------------------------------------
# --- "Q"

#------------------------------------------------------------------------------
# --- "R"
requests==2.28.2

# -----------------------------------------------------------------------------
# --- "S"

#------------------------------------------------------------------------------
# --- "T"

#------------------------------------------------------------------------------
# --- "U"

#------------------------------------------------------------------------------
# --- "V"

#------------------------------------------------------------------------------
# --- "W"

#------------------------------------------------------------------------------
# --- "X"

#------------------------------------------------------------------------------
# --- "Y"

#------------------------------------------------------------------------------
# --- "Z"
